+++ util/win/exception_handler_server.cc
@@ -347,7 +347,7 @@ std::wstring ExceptionHandlerServer::CreatePipe() {
 void ExceptionHandlerServer::Run(Delegate* delegate) {
   uint64_t shutdown_token = base::RandUint64();
   ScopedKernelHANDLE thread_handles[kPipeInstances];
-  for (int i = 0; i < arraysize(thread_handles); ++i) {
+  for (size_t i = 0; i < arraysize(thread_handles); ++i) {
     HANDLE pipe;
     if (first_pipe_instance_.is_valid()) {
       pipe = first_pipe_instance_.release();
@@ -399,7 +399,7 @@ void ExceptionHandlerServer::Run(Delegate* delegate) {
   }
 
   // Signal to the named pipe instances that they should terminate.
-  for (int i = 0; i < arraysize(thread_handles); ++i) {
+  for (size_t i = 0; i < arraysize(thread_handles); ++i) {
     ClientToServerMessage message;
     memset(&message, 0, sizeof(message));
     message.type = ClientToServerMessage::kShutdown;
